+++ src/io/segment/android/info/Device.java
@@ -8,6 +8,8 @@ import android.content.Context;
 
 public class Device implements Info<JSONObject> {
 
+	private SessionId sessionId = new SessionId();
+	
 	@Override
 	public String getKey() {
 		return "device";
@@ -21,7 +23,9 @@ public class Device implements Info<JSONObject> {
 		device.put("sdk", android.os.Build.VERSION.SDK_INT);
 		device.put("release",  android.os.Build.VERSION.RELEASE);
 		device.put("brand", android.os.Build.BRAND);
+		device.put("model", android.os.Build.MODEL);
 		device.put("manufacturer",  android.os.Build.MANUFACTURER);
+		device.put("id",  sessionId.get(context));
 		
 		return device;	
 	}